some documentation was added
