+++ Sources/RickAndMortySwiftApi/RAMEpisodes/RAMEpisodeModel.swift
@@ -2,14 +2,16 @@ import Foundation
 
 /**
 Rick and Morty Episode model
-- Parameters:
-    - id: The id of the episode (Int?). 
-    - name: The name of the episode (String?).
-    - airDate: The air date of the episode (String?).
-    - episode: The code of the episode, ex. "S01E01" (String?).
-    - charactes: List of characters who have been seen in the episode ([String]?).
-    - url: Link to the episode's own endpoint (String?).
-    - created: Time at which the episode was created in the database, ex. "2017-11-10T12:56:33.798Z" (String?).
+```
+// Parameters:
+//    id        : Int?      - The id of the episode. 
+//    name      : String?   - The name of the episode.
+//    airDate   : String?   - The air date of the episode.
+//    episode   : String?   - The code of the episode, ex. "S01E01".
+//    charactes : [String]? - List of characters who have been seen in the episode.
+//    url       : String?   - Link to the episode's own endpoint.
+//    created   : String?   - Time at which the episode was created in the database, ex. "2017-11-10T12:56:33.798Z" .
+```
 */
 public struct RAMEpisodeModel: Codable {
